refactor: extract payment gateway class
